# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: developers@givechariot.com
    name: Chariot Development Team
    url: https://www.givechariot.com/contact
  description: The Chariot Payments REST API.
  termsOfService: https://www.givechariot.com/legal-nonprofit
  title: Chariot Payments API
  version: v1
servers:
- description: Staging
  url: https://devapi.givechariot.com
- description: Sandbox
  url: https://sandboxapi.givechariot.com
- description: Production
  url: https://api.givechariot.com
security:
- bearerAuth: []
paths:
...
  /transfers:
    get:
      description: |
        List transfers.
        This endpoint is used to list transfers.
      operationId: listTransfers
      parameters:
      - description: The unique identifier for the account
        explode: true
        in: query
        name: account_id
        required: true
        schema:
          type: string
        style: form
      - description: Limit the size of the list that is returned. The default (and
          maximum) is 100 objects.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: "The cursor to use for pagination. If not set, the first page\
          \ of results will be returned."
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferList'
          description: The transfers were retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: List transfers
      tags:
      - transfers
    post:
      description: |
        Transfer funds between accounts.
        This endpoint is used to move money into and out of the Chariot account from an external bank account.
      operationId: transferFunds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
        description: The transfer to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
          description: The transfer was created
          headers:
            Location:
              $ref: '#/components/headers/Location'
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Transfer funds
      tags:
      - transfers
  /transfers/{id}:
    get:
      description: |
        Get a transfer by its unique identifier.
        This endpoint is used to retrieve a transfer by its unique identifier.
      operationId: getTransfer
      parameters:
      - description: The unique identifier for the transfer
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
          description: The transfer was retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "404":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 404
                error: Not Found
                message: Resource is not found
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Retrieve a transfer
      tags:
      - transfers
...
components:
  headers:
    X-Request-Id:
      description: The unique identifier for the request
      explode: false
      schema:
        type: string
      style: simple
...
    Location:
      description: The URI reference of the created object.
      explode: false
      schema:
        format: uri
        type: string
      style: simple
...
  schemas:
...
    Pagination:
      description: The paging information
      example:
        cursors:
          before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
        total: 100
      properties:
        cursors:
          $ref: '#/components/schemas/Pagination_cursors'
        total:
          description: The total number of objects
          example: 100
          type: integer
...
    Transfer:
      description: A transfer moves funds between accounts using ACH.
      example:
        transaction_id: transaction_uyrp7fld2ium70oa7oi
        account_number: "1234567890"
        amount: 10000
        funding: checking
        account_id: "1234567890"
        description: Transfer to FBO Account
        created_at: 2000-01-23T04:56:07.000+00:00
        id: "1234567890"
        routing_number: "123456789"
        status: pending
      properties:
        id:
          description: The unique identifier for the transfer
          example: "1234567890"
          readOnly: true
          type: string
        account_id:
          description: The identifier for the Chariot account that will send the transfer.
          example: "1234567890"
          type: string
        description:
          description: |
            A description of the transfer.
            Remittance information for the transfer that will show on the recipient's bank statement.
          example: Transfer to FBO Account
          type: string
        amount:
          description: The transfer amount in USD cents. A positive amount originates
            a credit transfer pushing funds to the receiving account. A negative amount
            originates a debit transfer pulling funds from the receiving account.
          example: 10000
          format: int64
          type: integer
        account_number:
          description: The account number for the destination account.
          example: "1234567890"
          type: string
        routing_number:
          description: The American Bankers' Association (ABA) Routing Transit Number
            (RTN) for the destination account.
          example: "123456789"
          type: string
        funding:
          description: The type of the account to which the transfer will be sent.
          enum:
          - checking
          - savings
          example: checking
          type: string
        transaction_id:
          description: The unique identifier for the transaction funding the transfer.
          example: transaction_uyrp7fld2ium70oa7oi
          readOnly: true
          type: string
        status:
          description: The status of the transfer
          enum:
          - canceled
          - pending
          - submitted
          - returned
          - rejected
          example: pending
          type: string
        created_at:
          description: The date and time the transfer was created
          format: date-time
          readOnly: true
          type: string
      required:
      - account_id
      - account_number
      - amount
      - description
      - routing_number
      type: object
    TransferList:
      description: A list of transfers
      example:
        data:
        - transaction_id: transaction_uyrp7fld2ium70oa7oi
          account_number: "1234567890"
          amount: 10000
          funding: checking
          account_id: "1234567890"
          description: Transfer to FBO Account
          created_at: 2000-01-23T04:56:07.000+00:00
          id: "1234567890"
          routing_number: "123456789"
          status: pending
        - transaction_id: transaction_uyrp7fld2ium70oa7oi
          account_number: "1234567890"
          amount: 10000
          funding: checking
          account_id: "1234567890"
          description: Transfer to FBO Account
          created_at: 2000-01-23T04:56:07.000+00:00
          id: "1234567890"
          routing_number: "123456789"
          status: pending
        paging:
          cursors:
            before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
            after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          total: 100
      properties:
        data:
          description: The list of transfers
          items:
            $ref: '#/components/schemas/Transfer'
          type: array
        paging:
          $ref: '#/components/schemas/Pagination'
    ApiError:
      description: An API error
      properties:
        timestamp:
          description: The date and time the error occurred
          format: date-time
          readOnly: true
          type: string
        status:
          description: The HTTP status code
          readOnly: true
          type: integer
        error:
          description: The HTTP status message
          readOnly: true
          type: string
        message:
          description: The error message
          readOnly: true
          type: string
      type: object
...
    Pagination_cursors:
      description: The cursors for the next and previous pages
      example:
        before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
        after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
      properties:
        before:
          description: The cursor for the previous page
          example: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          type: string
        after:
          description: The cursor for the next page
          example: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: Opaque
      scheme: bearer
      type: http